+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/)
 with respect to its command line interface and HTTP interface.
 
+## [Unreleased](//github.com/opentable/sous/compare/0.5.53...HEAD)
+### Added
+* All: a Schedule field on Manifests, which should publish to Singularity to allow for scheduled tasks.
+
 ## [0.5.53](//github.com/opentable/sous/compare/0.5.52...0.5.53)
 
 ### Fixed